JSONRPC 2.0 - Rename file that holds the registration functions to be exposed as public handlers.
