+++ src/traffic_server/Makefile.inc
@@ -55,8 +55,8 @@ traffic_server_traffic_server_SOURCES = \
 	traffic_server/InkAPI.cc \
 	traffic_server/InkIOCoreAPI.cc \
 	traffic_server/SocksProxy.cc \
-	traffic_server/RPCPubHandlers.cc \
-	traffic_server/RPCPubHandlers.h \
+	traffic_server/RpcAdminPubHandlers.cc \
+	traffic_server/RpcAdminPubHandlers.h \
 	shared/overridable_txn_vars.cc \
 	traffic_server/traffic_server.cc
